+++ toolproof/src/definitions/browser/mod.rs
@@ -1,13 +1,21 @@
+use std::path::PathBuf;
 use std::sync::Arc;
 
 use async_trait::async_trait;
+use chromiumoxide::cdp::browser_protocol::page::{
+    CaptureScreenshotFormat, CaptureScreenshotParams,
+};
 use chromiumoxide::cdp::browser_protocol::target::CreateTargetParams;
 use chromiumoxide::error::CdpError;
+use chromiumoxide::handler::viewport::Viewport;
+use chromiumoxide::page::ScreenshotParams;
 use futures::StreamExt;
 use tokio::task::JoinHandle;
 
 use crate::civilization::Civilization;
-use crate::errors::{ToolproofInputError, ToolproofInternalError, ToolproofStepError};
+use crate::errors::{
+    ToolproofInputError, ToolproofInternalError, ToolproofStepError, ToolproofTestFailure,
+};
 use crate::options::ToolproofParams;
 
 use super::{SegmentArgs, ToolproofInstruction, ToolproofRetriever};
@@ -34,13 +42,50 @@ async fn try_launch_browser(mut max: usize) -> (Browser, chromiumoxide::Handler)
     let mut launch = Err(CdpError::NotFound);
     while launch.is_err() && max > 0 {
         max -= 1;
-        launch = Browser::launch(BrowserConfig::builder().build().unwrap()).await;
+        launch = Browser::launch(
+            BrowserConfig::builder()
+                .headless_mode(chromiumoxide::browser::HeadlessMode::New)
+                .viewport(Some(Viewport {
+                    width: 1600,
+                    height: 900,
+                    device_scale_factor: Some(2.0),
+                    emulating_mobile: false,
+                    is_landscape: true,
+                    has_touch: false,
+                }))
+                .build()
+                .unwrap(),
+        )
+        .await;
     }
     match launch {
         Ok(res) => res,
         Err(e) => {
             panic!("Failed to launch browser due to error: {e}");
         }
+    }
+}
+
+fn chrome_image_format(filepath: &PathBuf) -> Result<CaptureScreenshotFormat, ToolproofStepError> {
+    match filepath.extension() {
+        Some(ext) => {
+            let ext = ext.to_string_lossy().to_lowercase();
+            match ext.as_str() {
+                "png" => Ok(CaptureScreenshotFormat::Png),
+                "webp" => Ok(CaptureScreenshotFormat::Webp),
+                "jpg" | "jpeg" => Ok(CaptureScreenshotFormat::Jpeg),
+                _ => Err(ToolproofStepError::External(
+                    ToolproofInputError::StepRequirementsNotMet {
+                        reason: "Image file extension must be png, webp, jpeg, or jpg".to_string(),
+                    },
+                )),
+            }
+        }
+        None => Err(ToolproofStepError::External(
+            ToolproofInputError::StepRequirementsNotMet {
+                reason: "Image file path must have an extension".to_string(),
+            },
+        )),
     }
 }
 
@@ -143,6 +188,63 @@ impl BrowserWindow {
                 .evaluate_script(harnessed(script))
                 .await
                 .map_err(|inner| ToolproofStepError::Internal(inner.into())),
+        }
+    }
+
+    async fn screenshot_page(&self, filepath: PathBuf) -> Result<(), ToolproofStepError> {
+        match self {
+            BrowserWindow::Chrome(page) => {
+                let image_format = chrome_image_format(&filepath)?;
+
+                page.save_screenshot(
+                    ScreenshotParams {
+                        cdp_params: CaptureScreenshotParams {
+                            format: Some(image_format),
+                            ..CaptureScreenshotParams::default()
+                        },
+                        full_page: Some(false),
+                        omit_background: Some(false),
+                    },
+                    filepath,
+                )
+                .await
+                .map(|_| ())
+                .map_err(|e| ToolproofStepError::Internal(e.into()))
+            }
+            BrowserWindow::Pagebrowse(_) => Err(ToolproofStepError::Internal(
+                ToolproofInternalError::Custom {
+                    msg: "Screenshots not yet implemented for Pagebrowse".to_string(),
+                },
+            )),
+        }
+    }
+
+    async fn screenshot_element(
+        &self,
+        selector: &str,
+        filepath: PathBuf,
+    ) -> Result<(), ToolproofStepError> {
+        match self {
+            BrowserWindow::Chrome(page) => {
+                let image_format = chrome_image_format(&filepath)?;
+
+                let element = page.find_element(selector).await.map_err(|e| {
+                    ToolproofStepError::Assertion(ToolproofTestFailure::Custom {
+                        msg: format!("Element {selector} could not be screenshot: {e}"),
+                    })
+                })?;
+
+                element
+                    .save_screenshot(image_format, filepath)
+                    .await
+                    .map(|_| ())
+                    .map_err(|e| ToolproofStepError::Internal(e.into()))
+            }
+            BrowserWindow::Pagebrowse(_) => Err(ToolproofStepError::Internal(
+                ToolproofInternalError::Custom {
+                    msg: "Screenshots not yet implemented for Pagebrowse".to_string(),
+                },
+            )),
         }
     }
 }
@@ -315,6 +417,79 @@ mod eval_js {
             civ: &mut Civilization,
         ) -> Result<serde_json::Value, ToolproofStepError> {
             eval_and_return_js("return toolproof_log_events[`ALL`];".to_string(), civ).await
+        }
+    }
+}
+
+mod screenshots {
+    use crate::errors::{ToolproofInternalError, ToolproofTestFailure};
+
+    use super::*;
+
+    pub struct ScreenshotViewport;
+
+    inventory::submit! {
+        &ScreenshotViewport as &dyn ToolproofInstruction
+    }
+
+    #[async_trait]
+    impl ToolproofInstruction for ScreenshotViewport {
+        fn segments(&self) -> &'static str {
+            "In my browser, I screenshot the viewport to {filepath}"
+        }
+
+        async fn run(
+            &self,
+            args: &SegmentArgs<'_>,
+            civ: &mut Civilization,
+        ) -> Result<(), ToolproofStepError> {
+            let filepath = args.get_string("filepath")?;
+            let resolved_path = civ.tmp_file_path(&filepath);
+            civ.ensure_path(&resolved_path);
+
+            let Some(window) = civ.window.as_ref() else {
+                return Err(ToolproofStepError::External(
+                    ToolproofInputError::StepRequirementsNotMet {
+                        reason: "no page has been loaded into the browser for this test".into(),
+                    },
+                ));
+            };
+
+            window.screenshot_page(resolved_path).await
+        }
+    }
+
+    pub struct ScreenshotElement;
+
+    inventory::submit! {
+        &ScreenshotElement as &dyn ToolproofInstruction
+    }
+
+    #[async_trait]
+    impl ToolproofInstruction for ScreenshotElement {
+        fn segments(&self) -> &'static str {
+            "In my browser, I screenshot the element {selector} to {filepath}"
+        }
+
+        async fn run(
+            &self,
+            args: &SegmentArgs<'_>,
+            civ: &mut Civilization,
+        ) -> Result<(), ToolproofStepError> {
+            let selector = args.get_string("selector")?;
+            let filepath = args.get_string("filepath")?;
+            let resolved_path = civ.tmp_file_path(&filepath);
+            civ.ensure_path(&resolved_path);
+
+            let Some(window) = civ.window.as_ref() else {
+                return Err(ToolproofStepError::External(
+                    ToolproofInputError::StepRequirementsNotMet {
+                        reason: "no page has been loaded into the browser for this test".into(),
+                    },
+                ));
+            };
+
+            window.screenshot_element(&selector, resolved_path).await
         }
     }
 }